Updated versioning and added config.h
